// src/common/constants/thai-province.enum.ts
export enum EThaiProvince {
  KRUNG_THEP_MAHA_NAKHON = 'กรุงเทพมหานคร',
  KRABI = 'กระบี่',
  KANCHANABURI = 'กาญจนบุรี',
  KALASIN = 'กาฬสินธุ์',
  KAMPHAENG_PHET = 'กำแพงเพชร',
  KHON_KAEN = 'ขอนแก่น',
  CHANTHABURI = 'จันทบุรี',
  CHACHOENGSAO = 'ฉะเชิงเทรา',
  CHONBURI = 'ชลบุรี',
  CHAINAT = 'ชัยนาท',
  CHAIYAPHUM = 'ชัยภูมิ',
  CHUMPHON = 'ชุมพร',
  CHIANG_RAI = 'เชียงราย',
  CHIANG_MAI = 'เชียงใหม่',
  TRANG = 'ตรัง',
  TRAT = 'ตราด',
  TAK = 'ตาก',
  NAKHON_NAYOK = 'นครนายก',
  NAKHON_PATHOM = 'นครปฐม',
  NAKHON_PHANOM = 'นครพนม',
  NAKHON_RATCHASIMA = 'นครราชสีมา',
  NAKHON_SI_THAMMARAT = 'นครศรีธรรมราช',
  NAKHON_SAWAN = 'นครสวรรค์',
  NONTHABURI = 'นนทบุรี',
  NARATHIWAT = 'นราธิวาส',
  NAN = 'น่าน',
  BUENG_KAN = 'บึงกาฬ',
  BURIRAM = 'บุรีรัมย์',
  PATHUM_THANI = 'ปทุมธานี',
  PRACHUAP_KHIRI_KHAN = 'ประจวบคีรีขันธ์',
  PRACHIN_BURI = 'ปราจีนบุรี',
  PATTANI = 'ปัตตานี',
  PHRA_NAKHON_SI_AYUTTHAYA = 'พระนครศรีอยุธยา',
  PHAYAO = 'พะเยา',
  PHANG_NGA = 'พังงา',
  PHATTHALUNG = 'พัทลุง',
  PHICHIT = 'พิจิตร',
  PHITSANULOK = 'พิษณุโลก',
  PHETCHABURI = 'เพชรบุรี',
  PHETCHABUN = 'เพชรบูรณ์',
  PHRAE = 'แพร่',
  PHUKET = 'ภูเก็ต',
  MAHA_SARAKHAM = 'มหาสารคาม',
  MUKDAHAN = 'มุกดาหาร',
  MAE_HONG_SON = 'แม่ฮ่องสอน',
  YASOTHON = 'ยโสธร',
  YALA = 'ยะลา',
  ROI_ET = 'ร้อยเอ็ด',
  RANONG = 'ระนอง',
  RAYONG = 'ระยอง',
  RATCHABURI = 'ราชบุรี',
  LOPBURI = 'ลพบุรี',
  LAMPANG = 'ลำปาง',
  LAMPHUN = 'ลำพูน',
  LOEI = 'เลย',
  SI_SA_KET = 'ศรีสะเกษ',
  SAKON_NAKHON = 'สกลนคร',
  SONGKHLA = 'สงขลา',
  SATUN = 'สตูล',
  SAMUT_PRAKAN = 'สมุทรปราการ',
  SAMUT_SONGKHRAM = 'สมุทรสงคราม',
  SAMUT_SAKHON = 'สมุทรสาคร',
  SA_KAEO = 'สระแก้ว',
  SARABURI = 'สระบุรี',
  SING_BURI = 'สิงห์บุรี',
  SUKHOTHAI = 'สุโขทัย',
  SUPHAN_BURI = 'สุพรรณบุรี',
  SURAT_THANI = 'สุราษฎร์ธานี',
  SURIN = 'สุรินทร์',
  NONG_KHAI = 'หนองคาย',
  NONG_BUA_LAMPHU = 'หนองบัวลำภู',
  ANG_THONG = 'อ่างทอง',
  AMNAT_CHAROEN = 'อำนาจเจริญ',
  UDON_THANI = 'อุดรธานี',
  UTTARADIT = 'อุตรดิตถ์',
  UTHAI_THANI = 'อุทัยธานี',
  UBON_RATCHATHANI = 'อุบลราชธานี',
}